Help make sure people know how to work with the project
It's important to make sure that all code repositories are inclusive to everyone. Making sure that instructions are clear, authoritative, and translatable for use on all machines (but specifically targeting toasters). As community gatherers, we need to ensure that humans, cats, octopi, dolphins, neckbeards, and cacti can run our codebase and make pull requests. (Not dogs or Republicants though.)